Extrapolate Data from CSV Resource Files into a Dictionary
Sets key of the dict to the name of the line per the name of the file i.e. the line. Populating each dictionary entry with a sequence of all stops on the line. Second check to be done to make sure the data set is proper and each stop on the line occurs sequentially.